+++ src/shim.ts
@@ -1,8 +1,9 @@
 /* used when calling require('htmlvalidate'); */
 
 export { default as HtmlValidate } from "./htmlvalidate";
+export { AttributeData } from "./parser";
+export { Config, ConfigData, ConfigLoader } from "./config";
 export { DynamicValue } from "./dom/dynamic-value";
 export { Rule } from "./rule";
-export { TemplateExtractor } from "./transform/template";
 export { Source } from "./context";
-export { AttributeData } from "./parser";
+export { TemplateExtractor } from "./transform/template";